voice_assistant_starter/requirements.txt
feedparser==6.0.11
requests>=2.31.0
SpeechRecognition>=3.10.4
pyttsx3>=2.90
pyaudio>=0.2.14
dateparser>=1.2.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
